+++ src/application/k2.c
@@ -46,7 +46,7 @@ static void rps_client() {
     RPSResponse response;
     int rpstid = WhoIs("RPS");
     rps_setup(rpstid);
-    srand(timer_read_raw(TIMER3));
+    srand((unsigned int) timer_read_raw(TIMER3));
     int rounds = rand() % (RPS_MAX_ROUNDS + 1);
 #ifdef OUTPUT
     bwprintf(COM2, "RPS Client %d: Game Start. Want to play %d rounds\n\r", MyTid(), rounds);
