+++ src/kern/event.c
@@ -23,13 +23,13 @@ void event_await(int tid, int event) {
             queue_push(&await_queue[event], tid);
             break;
         default:
-            current_task->tf->r0 = -1;
+            current_task->tf->r0 = (uint32_t) -1;
             break;
     }
 }
 
 void event_handle() {
-    int return_value;
+    uint32_t return_value;
     int event = icu_read();
     switch (event) {
         case TC1UI_EVENT:
